[Rename] : OnBoarding으로 파일명 변경 및 캐시 파일 삭제
