+++ app/on-boarding/page.tsx
@@ -1,4 +1,4 @@
-import ScreenOnBoarding from '@/components/screen/onBoarding';
+import ScreenOnBoarding from '@/components/screen/OnBoarding';
 
 export default function onBoarding() {
 	return <ScreenOnBoarding />;
